fix(verify): widen u32 dims before multiply + #[must_use] DiffResult
Code-quality review on Task 15 flagged:

1. (correctness) a8.width() * a8.height() multiplies in u32 before the
   cast to u64, overflowing for images >4 gigapixels. Cheap to harden:
   widen each dimension to u64 first.
2. (forward-compat) Drop a return value silently when calling
   compare_images. Add #[must_use] to DiffResult so callers can't.
3. (binary-fixture hygiene) Add .gitattributes marking PNG/JPG/WebP/
   AVIF/KTX2 as binary so git won't text-diff them as fixtures grow.

clippy::double_must_use caught the redundant fn-level #[must_use] when
the return type is already #[must_use]; dropped it.

+++ crates/buiy_verify/src/visual.rs
@@ -3,6 +3,7 @@
 
 use image::{DynamicImage, GenericImageView};
 
+#[must_use]
 pub struct DiffResult {
     /// 0.0 = identical, 1.0 = totally different.
     pub score: f64,
@@ -21,7 +22,9 @@ pub fn compare_images(a: &DynamicImage, b: &DynamicImage) -> DiffResult {
     let a8 = a.to_rgba8();
     let b8 = b.to_rgba8();
     let mut accumulated = 0u64;
-    let pixels = (a8.width() * a8.height()) as u64;
+    // Widen u32 → u64 BEFORE multiplying. `width * height` in u32 overflows
+    // for images > 4 gigapixels (theoretical, but cheap to harden).
+    let pixels = a8.width() as u64 * a8.height() as u64;
     for (pa, pb) in a8.pixels().zip(b8.pixels()) {
         for ch in 0..4 {
             let d = pa[ch] as i32 - pb[ch] as i32;
